Refactor weather shopper page objects
