+++ page_objects/examples/weather_shopper_mobile_app/weather_shopper_cart_page.py
@@ -1,80 +1,56 @@
 """
-Page objects for cart page in Weathershopper application.
+This class models the cart page in Weathershopper application.
 """
 # pylint: disable = W0212,E0401
-import re
 import conf.locators_conf as locators
 from utils.Wrapit import Wrapit
 from core_helpers.mobile_app_helper import Mobile_App_Helper
+from .weather_shopper_cart_objects import CartPageObjects
 
-class WeatherShopperCartPage(Mobile_App_Helper):
-    "Page object for the cart page in Weathershopper application."
-    @Wrapit._exceptionHandler
-    @Wrapit._screenshot
-    def change_quantity(self, item_to_change, quantity=2):
-        "This method is to change the quantity of an item in the cart"
-        result_flag = self.set_text(locators.edit_quantity.format(item_to_change), quantity)
-        self.conditional_write(result_flag,
-            positive=f"Successfully changed quantity of {item_to_change} to {quantity}",
-            negative=f"Failed to change quantity of {item_to_change} to {quantity}",
-            level='debug')
+class WeatherShopperCartPage(Mobile_App_Helper, CartPageObjects):
+    "Page objects for the cart page in Weathershopper application."
+
+    def verify_cart_total(self, items):
+        "Verify cart total"
+        cart_total = self.get_cart_total()
+        item_prices = [item['price'] for item in items]
+        result_flag = self.verify_total(cart_total, item_prices)
+
         return result_flag
 
-    @Wrapit._exceptionHandler
-    @Wrapit._screenshot
-    def checkout(self):
-        "This method is to go to the Checkout page"
-        result_flag = self.click_element(locators.checkout_button)
+    def change_quantity_and_verify(self, least_expensive_item,
+                                    most_expensive_item, quantity):
+        "Change quantity of item and verify cart total"
+        # Change quantity of least expensive item
+        result_flag = self.change_quantity(least_expensive_item['name'], quantity=quantity)
         self.conditional_write(result_flag,
-            positive="Successfully clicked on checkout button",
-            negative="Failed to click on checkout button",
-            level='debug')
-        self.switch_page("weathershopper payment page")
+                                positive="Successfully changed quantity of item",
+                                negative="Failed to change quantity of item")
+
+        # Refresh cart total
+        result_flag = self.refresh_total_amount()
+        self.conditional_write(result_flag,
+                                positive="Successfully refreshed total",
+                                negative="Failed to refresh total")
+
+        # Verify cart total after change in quantity
+        cart_total_after_change = self.get_cart_total()
+        item_prices = [least_expensive_item['price'] * quantity, most_expensive_item['price']]
+        result_flag = self.verify_total(cart_total_after_change, item_prices)
+
         return result_flag
 
-    @Wrapit._exceptionHandler
-    @Wrapit._screenshot
-    def delete_from_cart(self, item_to_delete):
-        "This method is to delete an item from the cart"
-        result_flag = self.click_element(locators.checkbox.format(item_to_delete))
-        result_flag &= self.click_element(locators.delete_from_cart_button)
+    def delete_item_and_verify(self, least_expensive_item, most_expensive_item, quantity):
+        "Delete item from cart and verify cart total"
+        # Delete item from cart
+        result_flag = self.delete_from_cart(most_expensive_item['name'])
         self.conditional_write(result_flag,
-            positive=f"Successfully deleted {item_to_delete} from cart",
-            negative="Failed to delete item from cart",
-            level='debug')
+                                positive="Successfully deleted item from cart",
+                                negative="Failed to delete item from cart")
+
+        # Verify cart total after deletion
+        cart_total_after_deletion = self.get_cart_total()
+        item_prices = [least_expensive_item['price'] * quantity]
+        result_flag = self.verify_total(cart_total_after_deletion, item_prices)
+
         return result_flag
-
-    @Wrapit._exceptionHandler
-    @Wrapit._screenshot
-    def get_cart_total(self):
-        "This method is to get the total price in the cart."
-        cart_total = self.get_text(locators.total_amount)
-        # Extracting the numeric part using regular expression
-        match = re.search(r'\d+\.\d+', cart_total.decode())
-        total_amount = float(match.group()) if match else None
-
-        if total_amount is None:
-            self.write("Total amount is None", level='debug')
-        else:
-            self.write(f"Total amount is {total_amount}", level='debug')
-        return total_amount
-
-    @Wrapit._exceptionHandler
-    @Wrapit._screenshot
-    def refresh_total_amount(self):
-        "This method is to refresh the total amount in the cart."
-        result_flag = self.click_element(locators.refresh_button)
-        self.conditional_write(result_flag,
-            positive="Successfully clicked on refresh button",
-            negative="Failed to click on refresh button",
-            level='debug')
-        return result_flag
-
-    @Wrapit._exceptionHandler
-    @Wrapit._screenshot
-    def verify_total(self, cart_total, cart_items):
-        "This method is to verify the total price in the cart."
-        if cart_total == sum(cart_items):
-            return True
-        return False
-
