+++ tests/test_weather_shopper_app.py
@@ -3,7 +3,6 @@ Automated test for Weather Shopper mobile application
 """
 
 # pylint: disable=E0401, C0413
-import secrets
 import time
 import os
 import sys
@@ -11,104 +10,6 @@ import pytest
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from page_objects.PageFactory import PageFactory
 import conf.weather_shopper_mobile_conf as conf
-
-def visit_product_page(test_mobile_obj, temperature):
-    "Visit the product page"
-    product = None
-    result_flag = False
-    if temperature < 19:
-        result_flag = test_mobile_obj.view_moisturizers()
-        product = "Moisturizers"
-
-    elif temperature > 32:
-        result_flag = test_mobile_obj.view_sunscreens()
-        product = "Sunscreens"
-
-    else:
-        skin_product = secrets.choice(['Moisturizers', 'Sunscreens'])
-        if skin_product == 'Moisturizers':
-            result_flag = test_mobile_obj.view_moisturizers()
-            product = "Moisturizers"
-        else:
-            result_flag = test_mobile_obj.view_sunscreens()
-            product = "Sunscreens"
-
-    return product, result_flag
-
-def get_items(test_mobile_obj):
-    "Get least and most expensive item from the page"
-
-    # Get all products from page
-    all_items = test_mobile_obj.get_all_products()
-
-    # Calculate least and most expensive item
-    least_expensive_item = test_mobile_obj.get_least_expensive_item(all_items)
-    most_expensive_item = test_mobile_obj.get_most_expensive_item(all_items)
-
-    return least_expensive_item, most_expensive_item
-
-def add_items_to_cart(test_mobile_obj, least_expensive_item, most_expensive_item):
-    "Add items to cart"
-    result_flag = test_mobile_obj.add_to_cart(least_expensive_item)
-
-    # Add most expensive item to cart
-    result_flag &= test_mobile_obj.add_to_cart(most_expensive_item)
-
-    return result_flag
-
-def view_cart(test_mobile_obj):
-    "View cart page"
-    result_flag = test_mobile_obj.view_cart()
-
-    return result_flag
-
-def verify_cart(test_mobile_obj, least_expensive_item, most_expensive_item):
-    "Verify cart total"
-    # Verify cart total
-    cart_total = test_mobile_obj.get_cart_total()
-    item_prices = [least_expensive_item['price'], most_expensive_item['price']]
-    result_flag = test_mobile_obj.verify_total(cart_total, item_prices)
-
-    return result_flag
-
-def change_quantity_and_verify(test_mobile_obj, least_expensive_item,
-                                most_expensive_item, quantity):
-    "Change quantity of item and verify cart total"
-    # Change quantity of least expensive item
-    result_flag = test_mobile_obj.change_quantity(least_expensive_item['name'], quantity=quantity)
-    test_mobile_obj.log_result(result_flag,
-                               positive="Successfully changed quantity of item",
-                               negative="Failed to change quantity of item",
-                               level = "critical")
-
-    # Refresh cart total
-    result_flag = test_mobile_obj.refresh_total_amount()
-    test_mobile_obj.log_result(result_flag,
-                               positive="Successfully refreshed total",
-                               negative="Failed to refresh total")
-
-    # Verify cart total after change in quantity
-    cart_total_after_change = test_mobile_obj.get_cart_total()
-    item_prices = [least_expensive_item['price'] * quantity, most_expensive_item['price']]
-    result_flag = test_mobile_obj.verify_total(cart_total_after_change, item_prices)
-
-    return result_flag
-
-def delete_item_and_verify(test_mobile_obj, least_expensive_item, most_expensive_item, quantity):
-    "Delete item from cart and verify cart total"
-    # Delete item from cart
-    result_flag = test_mobile_obj.delete_from_cart(most_expensive_item['name'])
-    test_mobile_obj.log_result(result_flag,
-                               positive="Successfully deleted item from cart",
-                               negative="Failed to delete item from cart",
-                               level = "critical")
-
-    # Verify cart total after deletion
-    cart_total_after_deletion = test_mobile_obj.get_cart_total()
-    item_prices = [least_expensive_item['price'] * quantity]
-    result_flag = test_mobile_obj.verify_total(cart_total_after_deletion, item_prices)
-
-    return result_flag
 
 @pytest.mark.MOBILE
 def test_weather_shopper_app(test_mobile_obj):
@@ -127,7 +28,7 @@ def test_weather_shopper_app(test_mobile_obj):
         temperature = test_mobile_obj.get_temperature()
 
         #Visit the product page
-        product,result_flag = visit_product_page(test_mobile_obj, temperature)
+        product,result_flag = test_mobile_obj.visit_product_page(temperature)
 
         if product == "Moisturizers":
             test_mobile_obj.log_result(result_flag,
@@ -140,32 +41,36 @@ def test_weather_shopper_app(test_mobile_obj):
                                     negative="Failed to visit sunscreens page",
                                     level="critical")
 
+        # Get all products from page
+        all_items = test_mobile_obj.get_all_products()
+
         #Get least and most expensive item from the page
-        least_expensive_item, most_expensive_item = get_items(test_mobile_obj)
+        least_expensive_item, most_expensive_item = test_mobile_obj.get_least_and_most_expensive_items(all_items)
+        items = [least_expensive_item, most_expensive_item]
 
         #Add items to cart
-        result_flag = add_items_to_cart(test_mobile_obj, least_expensive_item, most_expensive_item)
+        result_flag = test_mobile_obj.add_items_to_cart(items)
         test_mobile_obj.log_result(result_flag,
                                     positive="Successfully added items item to cart",
                                     negative="Failed to add one or mre items to the cart",
                                     level="critical")
 
         #View cart
-        result_flag = view_cart(test_mobile_obj)
+        result_flag = test_mobile_obj.view_cart()
         test_mobile_obj.log_result(result_flag,
                                     positive="Successfully viewed cart",
                                     negative="Failed to view cart",
                                     level="critical")
 
         #Verify cart total
-        result_flag = verify_cart(test_mobile_obj, least_expensive_item, most_expensive_item)
+        result_flag = test_mobile_obj.verify_cart_total(items)
         test_mobile_obj.log_result(result_flag,
                                 positive="Cart total is correct",
                                 negative="Total is incorrect")
 
         #Change quantity of item in cart and verify cart total
         quantity = 2
-        result_flag = change_quantity_and_verify(test_mobile_obj, least_expensive_item,
+        result_flag = test_mobile_obj.change_quantity_and_verify(least_expensive_item,
                         most_expensive_item, quantity)
 
         test_mobile_obj.log_result(result_flag,
@@ -173,7 +78,7 @@ def test_weather_shopper_app(test_mobile_obj):
                                 negative="Failed to change quantity of item")
 
         #Delete item from cart and verify cart total
-        result_flag = delete_item_and_verify(test_mobile_obj, least_expensive_item,
+        result_flag = test_mobile_obj.delete_item_and_verify(least_expensive_item,
                         most_expensive_item, quantity)
         test_mobile_obj.log_result(result_flag,
                             positive="Total after deletion is accurate",
